+++ Source/EventFlow/EventFlowOptions.cs
@@ -222,7 +222,7 @@ namespace EventFlow
             serviceRegistration.Register<ISagaStore, SagaAggregateStore>();
             serviceRegistration.Register<ISagaErrorHandler, SagaErrorHandler>();
             serviceRegistration.Register<IDispatchToSagas, DispatchToSagas>();
-#if NET451
+#if NET452
             serviceRegistration.Register<IMemoryCache, MemoryCache>(Lifetime.Singleton);
 #else
             serviceRegistration.Register<IMemoryCache, DictionaryMemoryCache>(Lifetime.Singleton);