+++ RELEASE_NOTES.md
@@ -18,6 +18,8 @@
   - `EventFlow.Sql`
   - `EventFlow.SQLite`
 * New: Added [SourceLink](https://github.com/dotnet/sourcelink) support
+* Fix: `DispatchToSagas.ProcessSagaAsync` use `EventId` instead of `SourceId` as `SourceId` 
+  for delivery of external event to AggregateSaga
 * Fix: `Identity<T>.NewComb()` now produces string values that doesn't cause
   too much index fragmentation in MSSQL string columns